Adjust pytest hook wrapper to show an reduced error message (#148)
To give a better overview, I added a custom long report in
`pytest_runtest_makereport` which is printed when the test fails. This
way the error type and the line number is still depicted when a test
fails but not the whole script. The path to the ORCA directory is shown
more prominently.